LayoutTests/platform/chromium/fast/events/onerror-no-constructor-expected.txt

CONSOLE MESSAGE: line 20: Uncaught PASS if this is reported as an uncaught exception.
